requirements.txt
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.4
CacheControl==0.12.5
certifi==2018.8.24
cffi==1.11.5
chainmap==1.0.2
chardet==3.0.4
click==6.7
colorama==0.4.0
combomethod==1.0.12
contextlib2==0.5.5
coverage==4.5.1
coveralls==1.5.1
cryptography==2.3.1
docopt==0.6.2
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-WTF==0.14.2
freezegun==0.3.10
gitdb2==2.0.5
GitPython==2.1.11
idna==2.7
iso8601==0.1.12
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mock==2.0.0
more-itertools==4.3.0
msgpack==0.5.6
multi-mechanize==1.2.0
nulltype==2.3.1
options==1.4.8
pathlib2==2.3.2
pbr==4.3.0
peewee==2.4.4
pep8==1.7.1
pluggy==0.7.1
psutil==5.4.7
py==1.7.0
pyasn1==0.4.4
pycparser==2.19
PyJWT==1.6.4
pyparsing==2.2.2
pytest==3.8.2
pytest-cov==2.6.0
pytest-random-order==0.8.0
python-dateutil==2.7.3
pytz==2018.9
pywin32==223
PyXB==1.2.6
rdflib==4.2.2
regex==2019.1.24
requests==2.19.1
requests-toolbelt==0.8.0
six==1.11.0
smmap2==2.0.5
Unidecode==1.0.23
urllib3==1.23
Werkzeug==0.14.1
WTForms==2.2.1
zipstream==1.1.4
